// src/components/RhTree/type.ts
export type IBaseNode = {
  /**
   * 节点id
   * @requires
   * @type string | number
   */
  id: string | number;
  /**
   * 节点名称
   * @requires
   * @type string
   */
  name: string;
  /**
   * 父节点id
   * @requires
   * @type string | number
   * @default null
   */
  parentId: string | number;
  /**
   * 数量
   */
  number?: string | number;
};

export interface ILeafNode extends DataNode, IBaseNode {
  /**
   * 节点是否可编辑
   * @type boolean
   */
  isEdit?: boolean;
  /**
   * 节点是否可新增
   * @type boolean
   */
  isCreate?: boolean;
  /**
   * 子节点数组
   * @type ILeafNode[]
   */
  children?: ILeafNode[];
  /**
   * 节点图标
   * @type undefined | string | ReactNode
   */
  icon?: undefined | string | ReactNode;
  /**
   * 节点类型，用于业务判断区分
   * @type string
   */
  type?: undefined | string;
}